docs: update icons example
