+++ packages/web3/src/icons/index.md
@@ -6,14 +6,26 @@ group:
 order: 1
 ---
 
+## How to use
+
+Before use icons, you need to install [@ant-design/web3-icons](https://www.npmjs.com/package/@ant-design/web3-icons) package:
+
+```sh
+# npm
+npm install @ant-design/web3-icons --save
+
+# yarn
+yarn add @ant-design/web3-icons
+
+# pnpm
+pnpm install @ant-design/web3-icons --save
+```
+
 ## List of icons
 
 <IconSearch></IconSearch>
 
-## Colorful style
+## Examples
 
-<code src="./demos/colorful.tsx"></code>
-
-## Filled style
-
-<code src="./demos/filled.tsx"></code>
+<!-- prettier-ignore -->
+<code src="./demos/basic.tsx">Basic</code>
