+++ packages/web3/src/icons/index.zh-CN.md
@@ -6,14 +6,26 @@ group:
 order: 1
 ---
 
+## 使用方法
+
+使用图标组件，你需要安装 [@ant-design/web3-icons](https://www.npmjs.com/package/@ant-design/web3-icons) 图标组件包：
+
+```sh
+# npm
+npm install @ant-design/web3-icons --save
+
+# yarn
+yarn add @ant-design/web3-icons
+
+# pnpm
+pnpm install @ant-design/web3-icons --save
+```
+
 ## 图标列表
 
 <IconSearch></IconSearch>
 
-## 彩色风格
+## 代码演示
 
-<code src="./demos/colorful.tsx"></code>
-
-## 单色风格
-
-<code src="./demos/filled.tsx"></code>
+<!-- prettier-ignore -->
+<code src="./demos/basic.tsx">基本用法</code>
